Initial commit of the Snap package latest-20190813
Initial commit of the files used to create Snap package version latest-20190813 (which is currently in the Ubuntu Snap Store). I created this separate project novnc-snap since the latest pull request to the snap/ directory in the novnc repo is still pending. This repo (novnc-snap) was created to have the latest version of the Snap package